+++ libs/icons/components/DashboardIcon.tsx
@@ -1,0 +1,4 @@
+import { Icon, IconProps } from '@chakra-ui/icons'
+import { MdPerson } from 'react-icons/md'
+
+export const DashboardIcon = (props: IconProps) => <Icon as={MdPerson} {...props} />
