+++ libs/icons/components/FilterIcon.tsx
@@ -1,0 +1,4 @@
+import { Icon, IconProps } from '@chakra-ui/icons'
+import { MdFilterList } from 'react-icons/md'
+
+export const FilterIcon = (props: IconProps) => <Icon as={MdFilterList} {...props} />
