+++ libs/icons/components/FullscreenExitIcon.tsx
@@ -1,0 +1,4 @@
+import { Icon, IconProps } from '@chakra-ui/icons'
+import { MdFullscreenExit } from 'react-icons/md'
+
+export const FullscreenExitIcon = (props: IconProps) => <Icon as={MdFullscreenExit} {...props} />
