+++ libs/icons/components/FullscreenIcon.tsx
@@ -1,0 +1,4 @@
+import { Icon, IconProps } from '@chakra-ui/icons'
+import { MdFullscreen } from 'react-icons/md'
+
+export const FullscreenIcon = (props: IconProps) => <Icon as={MdFullscreen} {...props} />
