+++ libs/icons/components/SellIcon.tsx
@@ -1,0 +1,4 @@
+import { Icon, IconProps } from '@chakra-ui/icons'
+import { MdSell } from 'react-icons/md'
+
+export const SellIcon = (props: IconProps) => <Icon as={MdSell} {...props} />
