+++ libs/icons/components/SortIcon.tsx
@@ -1,0 +1,4 @@
+import { Icon, IconProps } from '@chakra-ui/icons'
+import { MdSort } from 'react-icons/md'
+
+export const SortIcon = (props: IconProps) => <Icon as={MdSort} {...props} />
