+++ libs/icons/components/TransparencyIcon.tsx
@@ -1,0 +1,4 @@
+import { Icon, IconProps } from '@chakra-ui/icons'
+import { MdOutlineDashboard } from 'react-icons/md'
+
+export const TransparencyIcon = (props: IconProps) => <Icon as={MdOutlineDashboard} {...props} />
