+++ libs/icons/components/index.ts
@@ -1,5 +1,4 @@
 export { AirdropIcon } from './AirdropIcon'
-export * from './ReactIcons'
 export { DiscountIcon } from './DiscountIcon'
 export { DownIcon } from './DownIcon'
 export { ExpandArrowIcon } from './ExpandArrowIcon'
@@ -19,3 +18,10 @@ export { SwapIcon } from './SwapIcon'
 export { TransactionIcon } from './TransactionIcon'
 export { UnlockedIcon } from './UnlockedIcon'
 export { DiscordIcon } from './DiscordIcon'
+export { DashboardIcon } from './DashboardIcon'
+export { FilterIcon } from './FilterIcon'
+export { FullscreenIcon } from './FullscreenIcon'
+export { FullscreenExitIcon } from './FullscreenExitIcon'
+export { SellIcon } from './SellIcon'
+export { SortIcon } from './SortIcon'
+export { TransparencyIcon } from './TransparencyIcon'
